Restructured test directory layout. Seperated concerns for AddMemberToTeam, JoinTeam, CreateMembershipRequestToTeam Created messages directory to handle SlackMessage construction for commands and events.
