+++ src/gluon/commands/member/AddSlackDetails.ts
