+++ src/gluon/commands/member/OnboardMember.ts
@@ -9,9 +9,8 @@ import {
     Parameter,
     Tags,
 } from "@atomist/automation-client";
-import {buttonForCommand} from "@atomist/automation-client/spi/message/MessageClient";
-import {SlackMessage, url} from "@atomist/slack-messages";
 import {QMConfig} from "../../../config/QMConfig";
+import {OnboardMemberMessages} from "../../messages/member/OnboardMemberMessages";
 import {GluonService} from "../../services/gluon/GluonService";
 import {
     handleQMError,
@@ -19,8 +18,6 @@ import {
     ResponderMessageClient,
 } from "../../util/shared/Error";
 import {isSuccessCode} from "../../util/shared/Http";
-import {CreateTeam} from "../team/CreateTeam";
-import {JoinTeam} from "../team/JoinTeam";
 
 @CommandHandler("Onboard a new team member", QMConfig.subatomic.commandPrefix + " onboard me")
 @Tags("subatomic", "slack", "member")
@@ -94,39 +91,4 @@ export class OnboardMember implements HandleCommand<HandlerResult> {
         const messageClient = new ResponderMessageClient(ctx);
         return await handleQMError(messageClient, error);
     }
-}
-
-export class OnboardMemberMessages {
-    public presentTeamCreationAndApplicationOptions(firstName: string): SlackMessage {
-        const text: string = `
-Welcome to the Subatomic environment *${firstName}*!
-Next steps are to either join an existing team or create a new one.
-                `;
-
-        return {
-            text,
-            attachments: [{
-                fallback: "Welcome to the Subatomic environment",
-                footer: `For more information, please read the ${this.docs()}`,
-                thumb_url: "https://raw.githubusercontent.com/absa-subatomic/subatomic-documentation/gh-pages/images/subatomic-logo-colour.png",
-                actions: [
-                    buttonForCommand(
-                        {
-                            text: "Apply to join a team",
-                            style: "primary",
-                        },
-                        new JoinTeam()),
-                    buttonForCommand(
-                        {text: "Create a new team"},
-                        new CreateTeam()),
-                ],
-            }],
-        };
-    }
-
-    public docs(): string {
-        return `${url(`${QMConfig.subatomic.docs.baseUrl}/quantum-mechanic/command-reference#joinTeam`,
-            "documentation")}`;
-    }
-
 }
