+++ test/gluon/commands/team/CreateTeamTest.ts
@@ -2,9 +2,9 @@ import "mocha";
 import * as assert from "power-assert";
 const MockAdapter = require("axios-mock-adapter");
 import axios from "axios";
-import {QMConfig} from "../../../src/config/QMConfig";
-import {CreateTeam} from "../../../src/gluon/commands/team/CreateTeam";
-import {TestMessageClient} from "../TestMessageClient";
+import {QMConfig} from "../../../../src/config/QMConfig";
+import {CreateTeam} from "../../../../src/gluon/commands/team/CreateTeam";
+import {TestMessageClient} from "../../TestMessageClient";
 
 describe("Create Team test", () => {
     it("should create team", async () => {
