+++ test/gluon/events/team/MembershipRequestClosedTest.ts
@@ -2,10 +2,10 @@ import "mocha";
 import * as assert from "power-assert";
 const MockAdapter = require("axios-mock-adapter");
 import axios from "axios";
-import {QMConfig} from "../../../src/config/QMConfig";
-import {MembershipRequestClosed} from "../../../src/gluon/events/team/MembershipRequestClosed";
-import {TestGraphClient} from "../TestGraphClient";
-import {TestMessageClient} from "../TestMessageClient";
+import {QMConfig} from "../../../../src/config/QMConfig";
+import {MembershipRequestClosed} from "../../../../src/gluon/events/team/MembershipRequestClosed";
+import {TestGraphClient} from "../../TestGraphClient";
+import {TestMessageClient} from "../../TestMessageClient";
 
 describe("Close a membership request", () => {
     it("should approve team member", async () => {
